feat: refactor form steps for improved UX flow - Move deck name to step 2 (conditional on custom/AI deck types) - Integrate audio settings into step 3 (Content & Audio) - Consolidate API settings into step 5 (Model & AI Settings) - Remove redundant ApiConfigSection component - Improve step labeling and user guidance - Maintain backward compatibility with all form data
